# ironflow-sdk/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountState": {
        "description": "Derived state of an account, for display.",
        "enum": ["ok", "near_limit", "limited", "token_invalid", "never_used"],
        "type": "string"
      },
...
      "AccountWindowResponse": {
        "description": "One usage window of an account.",
        "properties": {
          "model_scope": {"description": "Model family the window applies to, `null` for every model.", "nullable": true, "type": "string"},
          "observed_at": {"description": "When the window was observed.", "format": "date-time", "type": "string"},
          "resets_at": {"description": "When the window resets.", "format": "date-time", "nullable": true, "type": "string"},
          "status": {"$ref": "#/components/schemas/AccountWindowStatus"},
          "utilization": {"description": "Fraction used, `0.0..=1.0`.", "format": "double", "type": "number"},
          "window": {"description": "Window name (`five_hour`, `seven_day`).", "type": "string"}
        },
        "required": ["window", "utilization", "status", "observed_at"],
        "type": "object"
      },
      "AccountWindowStatus": {
        "description": "Status of a usage window, as reported by the provider.",
        "enum": ["allowed", "allowed_warning", "rejected"],
        "type": "string"
      },
...
      "ProviderAccountResponse": {
        "description": "A Provider Account, without its credential.",
        "properties": {
          "alert_threshold": {"description": "Alert threshold.", "format": "double", "type": "number"},
          "auth_failed_at": {"description": "When the provider last rejected the credential.", "format": "date-time", "nullable": true, "type": "string"},
          "created_at": {"description": "Creation timestamp.", "format": "date-time", "type": "string"},
          "created_by": {"description": "Creating user.", "format": "uuid", "nullable": true, "type": "string"},
          "display_name": {"description": "Human-readable name.", "type": "string"},
          "enabled": {"description": "Whether the account may be selected.", "type": "boolean"},
          "expires_at": {"description": "When the credential expires.", "format": "date-time", "type": "string"},
          "id": {"description": "Account ID.", "format": "uuid", "type": "string"},
          "kind": {"description": "Account kind.", "type": "string"},
          "max_concurrency": {"description": "Maximum concurrent steps, `null` for unlimited.", "format": "int32", "minimum": 0, "nullable": true, "type": "integer"},
          "name": {"description": "Unique slug.", "type": "string"},
          "plan": {"description": "Subscription plan.", "nullable": true, "type": "string"},
          "priority": {"description": "Priority, lower is preferred.", "format": "int32", "type": "integer"},
          "state": {"$ref": "#/components/schemas/AccountState"},
          "tags": {"description": "Tags.", "items": {"type": "string"}, "type": "array"},
          "updated_at": {"description": "Last update timestamp.", "format": "date-time", "type": "string"},
          "windows": {"description": "Latest windows.", "items": {"$ref": "#/components/schemas/AccountWindowResponse"}, "type": "array"}
        },
        "required": ["id", "name", "display_name", "kind", "enabled", "priority", "tags", "alert_threshold", "expires_at", "created_at", "updated_at", "state", "windows"],
        "type": "object"
      },
...
      "UpdateProviderAccountRequest": {
        "additionalProperties": false,
        "description": "Request body to update a Provider Account. All fields optional; `name`\nand `kind` are immutable.",
        "properties": {
          "alert_threshold": {"description": "New alert threshold in `(0, 1]`.", "format": "double", "nullable": true, "type": "number"},
          "display_name": {"description": "New display name.", "nullable": true, "type": "string"},
          "enabled": {"description": "Enable or disable.", "nullable": true, "type": "boolean"},
          "expires_at": {"description": "New expiry of the credential.", "format": "date-time", "nullable": true, "type": "string"},
          "max_concurrency": {"description": "New max concurrency; `null` removes the limit.", "format": "int32", "minimum": 0, "nullable": true, "type": "integer"},
          "plan": {"description": "New plan; `null` clears it.", "nullable": true, "type": "string"},
          "priority": {"description": "New priority.", "format": "int32", "nullable": true, "type": "integer"},
          "tags": {"description": "New tags (replaces the list).", "items": {"type": "string"}, "nullable": true, "type": "array"},
          "token": {"description": "Replacement credential, write-only. Checked against the provider.", "nullable": true, "type": "string"}
        },
        "type": "object"
      },
...
      "AuditLogEntry": {
        "description": "A persisted audit log entry capturing a domain event.\n\nEach entry records the full event payload as JSON alongside\ndenormalized contextual IDs for efficient filtering.\n\n# Examples\n\n```\nuse ironflow_store::entities::{AuditLogEntry, EventKind};\nuse uuid::Uuid;\nuse chrono::Utc;\nuse serde_json::json;\n\nlet entry = AuditLogEntry {\n    id: Uuid::now_v7(),\n    event_type: EventKind::RunStatusChanged,\n    payload: json!({\"run_id\": \"...\"}),\n    run_id: Some(Uuid::now_v7()),\n    step_id: None,\n    user_id: None,\n    created_at: Utc::now(),\n};\nassert_eq!(entry.event_type, EventKind::RunStatusChanged);\n```",
        "properties": {
          "created_at": {
            "description": "When the event was recorded.",
            "format": "date-time",
            "type": "string"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventKind",
            "description": "Event type."
          },
          "id": {
            "description": "Unique entry ID (UUID v7).",
            "format": "uuid",
            "type": "string"
          },
          "payload": {
            "description": "Full serialized event payload."
          },
          "run_id": {
            "description": "Associated run ID, if the event is run-scoped.",
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "step_id": {
            "description": "Associated step ID, if the event is step-scoped.",
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "user_id": {
            "description": "Associated user ID, if the event is user-scoped.",
            "format": "uuid",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "event_type",
          "payload",
          "created_at"
        ],
        "type": "object"
      },
...
      "EventKind": {
        "description": "Strongly-typed event kind matching domain event variants.\n\nSerializes to/from `snake_case` strings (e.g. `\"run_status_changed\"`).\n\n# Examples\n\n```\nuse ironflow_store::entities::EventKind;\n\nlet kind: EventKind = \"run_status_changed\".parse().unwrap();\nassert_eq!(kind, EventKind::RunStatusChanged);\nassert_eq!(kind.as_str(), \"run_status_changed\");\n```",
        "enum": [
          "run_created",
          "run_status_changed",
          "run_failed",
          "run_budget_exceeded",
          "step_completed",
          "step_failed",
          "approval_requested",
          "approval_granted",
          "approval_rejected",
          "approval_escalated",
          "log_line",
          "user_signed_in",
          "user_signed_up",
          "user_signed_out",
          "secrets_rotated",
          "retry_forced",
          "provider_account.updated",
          "provider_account.usage_updated"
        ],
        "type": "string"
      },
...
      "SecretResponse": {
        "description": "Response DTO for a secret (never exposes the value).",
        "properties": {
          "created_at": {
            "description": "Creation timestamp.",
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "description": "Secret ID.",
            "format": "uuid",
            "type": "string"
          },
          "key": {
            "description": "Secret key.",
            "type": "string"
          },
          "updated_at": {
            "description": "Last update timestamp.",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "UpdateSecretRequest": {
        "description": "Request body for updating a secret value.",
        "properties": {
          "value": {
            "description": "New secret value (plaintext, will be encrypted at rest).",
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Thomas Tartrau"
    },
    "description": "REST API for the ironflow workflow engine",
    "license": {
      "identifier": "MIT",
      "name": "MIT"
    },
    "title": "Ironflow REST API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/provider-accounts/{id}": {
      "delete": {
        "operationId": "delete_provider_account",
        "parameters": [{"description": "Account UUID or name", "in": "path", "name": "id", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "204": {"description": "Provider account deleted"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Provider account not found"}
        },
        "security": [{"Bearer": []}],
        "summary": "Delete a Provider Account and its credential. Admin only.",
        "tags": ["provider-accounts"]
      },
      "get": {
        "operationId": "get_provider_account",
        "parameters": [{"description": "Account UUID or name", "in": "path", "name": "id", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProviderAccountResponse"}}}, "description": "Provider account"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Provider account not found"}
        },
        "security": [{"Bearer": []}],
        "summary": "Get a Provider Account by UUID or name. Admin only.",
        "tags": ["provider-accounts"]
      },
      "patch": {
        "operationId": "update_provider_account",
        "parameters": [{"description": "Account UUID or name", "in": "path", "name": "id", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateProviderAccountRequest"}}}, "description": "Fields to change", "required": true},
        "responses": {
          "200": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProviderAccountResponse"}}}, "description": "Provider account updated"},
          "400": {"description": "Invalid input"},
          "401": {"description": "Unauthorized"},
          "403": {"description": "Forbidden"},
          "404": {"description": "Provider account not found"},
          "422": {"description": "Credential rejected"},
          "502": {"description": "Provider unreachable"}
        },
        "security": [{"Bearer": []}],
        "summary": "Update a Provider Account. Admin only.",
        "tags": ["provider-accounts"]
      }
    },
...
    "/api/v1/audit-logs": {
      "get": {
        "description": "Admin-only. Returns a paginated list of persisted domain events\nfor compliance review and post-mortem debugging.",
        "operationId": "list_audit_logs",
        "parameters": [
          {
            "description": "Filter by event type (e.g. `run_status_changed`).",
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/EventKind"
            }
          },
          {
            "description": "Filter by run ID.",
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Filter entries created at or after this timestamp.",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Filter entries created at or before this timestamp.",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "description": "Page number (1-based, default: 1).",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Items per page (default: 50, max: 100).",
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "format": "int32",
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditLogEntry"
                  },
                  "type": "array"
                }
              }
            },
            "description": "List of audit log entries with pagination"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin only"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List audit log entries with optional filtering and pagination.",
        "tags": [
          "audit"
        ]
      }
    },
...
    "/api/v1/secrets/{key}": {
      "delete": {
        "description": "# Errors\n\n- 403 if the caller is not an admin\n- 404 if the secret key does not exist",
        "operationId": "delete_secret",
        "parameters": [
          {
            "description": "Secret key",
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Secret deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Secret not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Delete a secret by key. Admin only.",
        "tags": [
          "secrets"
        ]
      },
      "put": {
        "description": "The key is taken from the URL path. The response never includes\nthe secret value.\n\n# Errors\n\n- 403 if the caller is not an admin\n- 404 if the secret key does not exist\n- 400 if the value is invalid",
        "operationId": "update_secret",
        "parameters": [
          {
            "description": "Secret key",
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSecretRequest"
              }
            }
          },
          "description": "New secret value",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecretResponse"
                }
              }
            },
            "description": "Secret updated"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Secret not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Update a secret's value. Admin only.",
        "tags": [
          "secrets"
        ]
      }
    },
...
  },
...
}